% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{addERRS}
\alias{addERRS}
\title{Add peak matching error windows using user-defined \emph{m/z} and \emph{rt} values}
\usage{
addERRS(dt, mz_err, rt_err)
}
\arguments{
\item{dt}{\code{data.frame}}

\item{mz_err}{\code{numeric} specifying error value for \emph{m/z}.}

\item{rt_err}{\code{numeric} specifying error value for \emph{rt}.}
}
\value{
Function returns original \code{data.frame} with additional columns 'mz_l', 'mz_h', 'rt_l' and 'rt_h'.
}
\description{
Function adds lower and higher values for \emph{m/z} and \emph{rt} values for each peak in the table.
}
