% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\name{adjustBATCH,massFlowTemplate-method}
\alias{adjustBATCH,massFlowTemplate-method}
\alias{adjustBATCH}
\title{adjustBATCH}
\usage{
\S4method{adjustBATCH}{massFlowTemplate}(
  object,
  out_dir = NULL,
  batch_end = NULL,
  batch_start = NULL,
  batch_next_metadata = NULL,
  batch_end_roi = NULL,
  batch_start_roi = NULL
)
}
\arguments{
\item{object}{\code{massFlowTemplate} class object.}

\item{out_dir}{\code{character} specifying desired directory for output.}

\item{batch_end}{\code{character} with filename of the last sample in batch No 1.}

\item{batch_start}{\code{character} with filename of the first sample in batch No 2.}

\item{batch_next_metadata}{\code{character} with path to metadata file for batch No 2.}

\item{batch_end_roi}{\code{character} with path to csv file with regions of integrations for the last sample in batch No 1.}

\item{batch_start_roi}{\code{character} with path to csv file with regions of integrations for the first sample in batch No 2.}
}
\value{
Method returns \code{massFlowTemplate} class object with updated template, such that \emph{rt} of the features are adjusted according to the observed batch-change effect.
}
\description{
Development mode.
Method adjusts retention time between two analytical batches using a set of features for which regions of integration (ROI) in \emph{mz}, \emph{rt} and \emph{into} domains are specified by the user.
}
