% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\name{alignPEAKS,massFlowTemplate-method}
\alias{alignPEAKS,massFlowTemplate-method}
\alias{alignPEAKS}
\title{Align peaks detected in LC-MS samples using spectral similarity comparison}
\usage{
\S4method{alignPEAKS}{massFlowTemplate}(object, out_dir = NULL, ncores = 2, cutoff = 0, write_int = FALSE)
}
\arguments{
\item{object}{\code{massFlowTemplate} class object, created by \code{buildTMP} constructor function.}

\item{out_dir}{\code{character} specifying desired directory for output.}

\item{ncores}{\code{numeric} for number of parallel workers to be used. Set 1 for serial implementation. Default set to 2.}

\item{cutoff}{\code{numeric} for spectra similarity score threshold, set to 0 by default.}

\item{write_int}{\code{logical} specifying whether a peak table with alignment results should be saved for every sample.
If TRUE, csv files will be written in the out_dir directory.
Default set to FALSE}
}
\value{
Method updates \code{\link{massFlowTemplate}} class object.
Slot @tmp is updated after each round of sample alignment.
Slot @data stores alignment results for each sample in the experiment.
Method writes alignment results to a csv file for each sample in the experiment.
}
\description{
Method aligns peaks across samples in LC-MS experiment using spectral similarity comparison.
To enable alignment, peaks originating from the same chemical compound were grouped into peak-groups, via function \code{\link{groupPEAKS}}.
}
\details{
Peaks are aligned across samples in their original acquisition order.
Template is list of all previously detected and aligned peaks.
For each peak in a sample, \code{alignPEAKS}:
\itemize{
\item Finds all template peaks within a \emph{m/z} and \emph{rt} window.
\item Identifies the true match by comparing the spectral similarity between the peak-group of the peak-of-interest and all matching template's peak-groups.
\item Merges the selected template's peak-group with the peak-group of the peak-of-interest.
\item Updates template's \emph{m/z} and \emph{rt} values for the matching peaks across the template and the sample.
}

Spectral similarity is measured by obtaining the cosine of the angle between two 2D vectors, representing each peak-group's \emph{m/z} and \emph{intensity} values.
}
\seealso{
\code{\link{buildTMP}}, \code{\link{validPEAKS}}, \code{\link{massFlowTemplate-class}}
}
