% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\docType{methods}
\name{alignSAMPLES,massFlowTemplate-method}
\alias{alignSAMPLES,massFlowTemplate-method}
\title{Align LC-MS experiment samples using groups of structurally-related peaks}
\usage{
\S4method{alignSAMPLES}{massFlowTemplate}(object)
}
\arguments{
\item{A}{\code{massFlowTemplate} class object, created by \emph{buildTMP} constructor function.}
}
\value{
A \code{massFlowTemplate} class object with updated \code{tmp}, \code{data} slots.
Method also writes intermediate alignment results for every sample in a csv file.
}
\description{
Merge peak-groups together across samples in the experiment using spectral similarity comparison.
}
\details{
A sample is aligned to the template, i.e. list of all previously detected peaks, by merging its peaks to the template peaks.
Matching peaks are found through \emph(m/z) and \emph(rt) windows. To identify true matches, the overall spectral similarity between a peak-group of the sample and all matching template peak-groups is compared.
Spectral similarity is measured by obtaining the cosine of the angle between two 2D vectors, representing each peak-group's \emph(m/z) and \emph(intensity) values. A peak-group of the sample is merged/grouped with the template's peak-group with which the highest cosine was obtained.
}
