% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowAnno.R
\docType{methods}
\name{annotateDS,massFlowAnno-method}
\alias{annotateDS,massFlowAnno-method}
\alias{annotateDS}
\title{Annotate dataset using chemical reference database.}
\usage{
\S4method{annotateDS}{massFlowAnno}(object, db_file = NULL,
  out_dir = NULL, mz_err = 0.01, rt_err = 10, ncores = 2)
}
\arguments{
\item{object}{\code{massFlowAnno} class object.}

\item{db_file}{\code{character} for absolute path to the csv file of the chemical reference database, built using \code{\link{buildDB}} function.}

\item{out_dir}{\code{character} specifying desired directory for output.}

\item{mz_err}{\code{numeric} specifying the window for peak matching in the MZ dimension. Default set to 0.01.}

\item{rt_err}{\code{numeric} specifying the window for peak matching in the RT dimension. Default set to 10 (sec).}

\item{ncores}{\code{numeric} for number of parallel workers to be used. Set 1 for serial implementation. Default set to 2.}
}
\value{
Method writes an updated dataset table with columns 'chemid', 'dbname', and 'similarity',
indicating matched chemical reference standards from the database.
Column 'similarity' indicates the stringth of spectral similarity between the PCS and the corresponding chemical compound, ranging from 0 to 1.
}
\description{
Method annotates a dataset using a chemical reference database.
Peak-groups, or Pseudo Chemical Spectra (PCS), in the dataset are compared with the peak-groups representing reference compounds in the database.
PCS are annotated using dot-product estimation between the spectra of the PCS and the corresponding chemical in the database.
}
\seealso{
\code{\link{buildDB}}, \code{\link{buildANNO}}
}
