% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-annotatePEAKS.R
\name{annotatePEAKS}
\alias{annotatePEAKS}
\title{Annotate dataset using a chemical reference database file}
\usage{
annotatePEAKS(dataset = NULL, database = NULL, out_dir = NULL,
  mz_err = 0.1, rt_err = 10, bins = 0.1, ncores = 2)
}
\arguments{
\item{dataset}{\code{character} for absolute path to the csv file of the dataset of interest.}

\item{database}{\code{character} for absolute path to the csv file of the chemical reference database, built using \code{\link{buildDB}} function.}

\item{out_dir}{\code{character} specifying desired directory for output.}

\item{mz_err}{\code{numeric} specifying the window for peak matching in the MZ dimension. Default set to 0.01.}

\item{rt_err}{\code{numeric} specifying the window for peak matching in the RT dimension. Default set to 10 (sec).}

\item{bins}{\code{numeric} defying step size used in peak-group spectra binning and vector generation. Step size represents MZ dimension. Default set to 0.01.}

\item{ncores}{\code{numeric} for number of parallel workers to be used. Set 1 for serial implementation. Default set to 2.}
}
\value{
Function writes an updated dataset table with columns 'db_peakid' and 'db_chemid', indicating matched chemical reference standards from the database.
Column 'cos' indicates the stringth of spectral similarity between the PCS and the corresponding chemical compound, ranging from 0 to 1.
}
\description{
Function annotates a dataset using a chemical reference database.
Peak-groups, or Pseudo Chemical Spectra (PCS), in the dataset are compared with the peak-groups representing reference compounds in the database.
PCS are annotated using dot-product estimation between the spectra of the PCS and the corresponding chemical in the database.
}
\seealso{
\code{\link{buildDB}}, \code{\link{alignPEAKS}}
}
