% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{buildANNO}
\alias{buildANNO}
\title{Build a \code{massFlowAnno} class object.}
\usage{
buildANNO(ds_file = NULL, meta_file = NULL, out_dir = NULL)
}
\arguments{
\item{ds_file}{\code{character} for absolute path to the csv file with final peak table obtained with \code{\link{fillPEAKS}}.}

\item{meta_file}{\code{character} for absolute path to the csv file, specifying samples filenames and their acquisition order.}

\item{out_dir}{\code{character} specifying desired directory for output.}
}
\value{
A \code{massFlowAnno} class object.
}
\description{
Build a \code{massFlowAnno} class object.
}
