% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-buildCLUSTS.R
\name{buildCLUSTS}
\alias{buildCLUSTS}
\title{Component assignment into chromatographic regions (clusters).}
\usage{
buildCLUSTS(fname, comp_table, out_dir)
}
\arguments{
\item{fname}{A \code{character} specifying datafile name.}

\item{comp_table}{A \code{DataFrame} class peak table with assigned components, created by \emph{buildCOMPS} function.}

\item{out_dir}{out_dir \code{character} object specifying directory where output data will be saved.}
}
\value{
Function returns an updated peak table. Column \code{"cls"} contains cluster ID for each peak.
Function also writes updated peak table to a text file in the specified directory.
}
\description{
Function assigns components of correlated peaks into chromatographic regions, called clusters.
}
