% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{buildDB}
\alias{buildDB}
\title{Build a chemical database template}
\usage{
buildDB(file = NULL)
}
\arguments{
\item{file}{\code{character} with absolute path to the the database template file (csv).}
}
\value{
Function returns a \code{massFlowDB} class object.
}
\description{
Function creates a \code{massFlowDB} class object, which will be used for chromatograpic peaks annotation and study samples alignment.
}
\examples{
## Build database template using a sample file, provided with the package
data_dir <- system.file("testdata", package = "massFlowR")
db_fname <- file.path(data_dir, "DBtemplate.csv")
db <- buildDB(file = db_fname)

}
\seealso{
\code{\link{massFlowDB}} class
}
