% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildGRAPH}
\alias{buildGRAPH}
\title{Build a correlation network of peaks}
\usage{
buildGRAPH(pkg_cor, cor_thr, plot = TRUE, title = NULL,
  out_dir = NULL)
}
\arguments{
\item{pkg_cor}{\code{data.frame} containing a row for each unique peak-peak pair. Columns 'from' and 'to' specify peak identifiers, column 'weight' represents correlation coefficient.}

\item{cor_thr}{\code{numeric} specifying a threshold for correlation coefficient below which peak pairs are omitted from the network.}

\item{plot}{\code{logical} whether to save a PNG of the built network.}

\item{title}{\code{character} specifying name for the plot, required if PLOT = TRUE.}

\item{out_dir}{\code{character} specifying absolute path to output directory, where a PNG would be written.}
}
\value{

}
\description{
Function build a correlation network of peaks using \emph{igraph} package functionality.
Only connections above the selected correlation threshold (\code{cor_thr}) will be retained in the network.
}
\details{
Function requires a data.frame with unique peak-peak pairs and correlation coefficients describing their relationship.
Such data frame can be built using \code{\link{getCORmat}} function.
This function is used to build correlation networks using coefficients that represent either EIC correlation of peaks in a single LC-MS sample, or inter-sample intensity correlation.
}
