% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{buildTMP}
\alias{buildTMP}
\title{Build a sample alignment and annotation template.}
\usage{
buildTMP(file = NULL, db = NULL, mz_err = 0.01, rt_err = 2,
  bins = 0.1, db_thrs = 0.5)
}
\arguments{
\item{file}{A \code{character} with path to the csv file, specifying samples filenames and their acquisition order.}

\item{db}{NULL, or a \code{massFlowDB} class object, built with \code{massFlowDB()} function.}

\item{mz_err}{A \code{numeric} specifying the window for peak matching in the MZ dimension. Default set to 0.01.}

\item{rt_err}{A \code{numeric} specifying the window for peak matching in the RT dimension. Default set to 2 (sec).}

\item{bins}{A \code{numeric} defying step size used in component's spectra binning and vector generation. Step size represents MZ dimension (default set to 0.1).}

\item{db_thrs}{A \code{numeric} specifying spectra similarity threshold (cosine) for first template generation with the database template (default set to 0.5).}
}
\value{
A \code{\link{massFlowTemplate}} class object.
}
\description{
Functions build \code{massFlowTemplate} class object, which stores study sample information.
}
