% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{buildVECTOR}
\alias{buildVECTOR}
\title{Build a vector for a peak-group using peaks \emph{m/z} and intensity values}
\usage{
buildVECTOR(spec, peaks)
}
\arguments{
\item{spec}{\code{data.frame} with columns 'mz', 'bin' and 'into'.
Contains a full range of \emph{m/z} values of all peak-groups that are being compared.}

\item{peaks}{\code{data.frame} with columns 'mz', 'bin' and 'into' for peaks in a single peak-group.}
}
\value{
Function returns a \code{numeric} vector representing peaks of a single peak-group.
}
\description{
Function builds a vector for a peak-group using peaks \emph{m/z} and intensity values.
Function is used to compare multiple peak-groups.
A vector is build for each peak-group separately using full \emph{m/z} range, which includes all peaks from all peak-groups being compared.
}
