% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowAnno.R
\name{checkANNOTATION,massFlowAnno-method}
\alias{checkANNOTATION,massFlowAnno-method}
\alias{checkANNOTATION}
\title{Check annotation results between selected database chemical compound and dataset pseudo chemical spectra.}
\usage{
\S4method{checkANNOTATION}{massFlowAnno}(object, chemid = NULL, pcs = NULL, out_dir = NULL)
}
\arguments{
\item{object}{\code{massFlowAnno} class object.}

\item{chemid}{\code{numeric} specifying one chemical compound id to look at.}

\item{pcs}{\code{numeric} specifying one pseudo chemical spectra to look at.}

\item{out_dir}{\code{character} specifying desired directory for output.}
}
\value{
Method returns two plots and write a csv file: (1) spectra of the selected chemical compound and pseudo chemical spectra from the dataset;
(2) retention time of the selected chemical compound and pseudo chemical spectra from the dataset; (3) table with the selected compound and pseudo chemical spectra.
}
\description{
Method returns summary plots of annotation between the selected chemical compound and pseudo chemical spectra in the dataset.
If out_dir is provided, generated plots are written as png files. Otherwise, plots are plotted on current graphical device.
\code{massFlowAnno} class object must be annotated using \code{\link{annotateDS}} method first.
}
