% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{checkFILE}
\alias{checkFILE}
\title{Check and load file for peak alignment}
\usage{
checkFILE(file = NULL)
}
\arguments{
\item{file}{\code{character} with absolute path to peak-group table.}
}
\value{
Function returns a \code{data.frame} of the corresponding file, if table was written correctly.
}
\description{
Internal function used within method \code{\link{alignPEAKS}}.
Function checks if file contains all neccessary columns and returns loaded \code{data.frame} if it is correct.
}
