% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\docType{methods}
\name{checkNEXT,massFlowTemplate-method}
\alias{checkNEXT,massFlowTemplate-method}
\title{Extract and check the filename of the sample to be processed next}
\usage{
\S4method{checkNEXT}{massFlowTemplate}(object)
}
\arguments{
\item{massFlowTemplate}{}
}
\value{

}
\description{
Extract and check the filename of the sample to be processed next
}
