% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-checkSPIKED.R
\name{checkSPIKED}
\alias{checkSPIKED}
\title{Title}
\usage{
checkSPIKED(fname, pks, raw, eic, spk, out_dir, prefilter, noise, rt_err = 3,
  mz_err = 0.005)
}
\arguments{
\item{mz_err}{}
}
\description{
Title
}
