% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-joinPEAKS_helpers.R
\name{checkTARGETpeak}
\alias{checkTARGETpeak}
\title{Check a given target peak and all of its matching peaks.}
\usage{
checkTARGETpeak(t_peakid, target_dt, mat_all)
}
\arguments{
\item{t_peakid}{\code{integer} with the current target peak's peakid.}

\item{target_dt}{\code{data.frame} with all peaks that belong to the target peak.}

\item{mat_all}{\code{data.frame} with all matching peaks.}
}
\value{

}
\description{
Function evaluates all of the matching peaks for a given target peak.
Function performs polynomial model fitting using RT values.
}
