% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowDB.R
\docType{methods}
\name{chemicals,massFlowDB-method}
\alias{chemicals,massFlowDB-method}
\title{Get the list of chemicals in the database}
\usage{
\S4method{chemicals}{massFlowDB}(object)
}
\arguments{
\item{object}{\code{massFlowDB} class object}
}
\description{
Get the list of chemicals in the database
}
