% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanPEAKS}
\alias{cleanPEAKS}
\title{Clean peak table from duplicating peaks}
\usage{
cleanPEAKS(rn, dt_unique, dt)
}
\arguments{
\item{rn}{\code{numeric} specifying a single row number (rowname) of the peak table.}

\item{dt_unique}{\code{data.frame} with unique peaks' \emph{m/z} and \emph{rt} values, with as many rows, as there are unique peaks.}

\item{dt}{\code{data.frame} object containing the peak table from which duplicated peaks should be removed.}
}
\value{
Function returns a single peak entry for the unique \emph{m/z} and \emph{rt} combination.
}
\description{
Function for every duplicated peak returns a single entry.
Peak duplication can appear because of:
\itemize{
\item artefacts in centWave peak table (identical mz/rt) (inside \code{\link{pickPEAKS}} function)
\item different datafile-of-interest peaks matching the same template peaks (identical mz/rt/peakid) (inside \code{\link{addDOI}} function)
}
}
