% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupPEAKS_helpers.R
\name{corEIC}
\alias{corEIC}
\title{Obtain EIC correlation between two peaks}
\usage{
corEIC(pair, eic)
}
\arguments{
\item{pair}{\code{matrix} with columns 'from' and 'to', specifying peak identifiers, which correspond to the indeces of a list of EICs.}

\item{eic}{\code{list} containing EIC for each picked-peak in a sample.}
}
\value{
Function returns EIC correlation coefficient.
}
\description{
Function performs extracted ion chromatogram (EIC) correlation of two peaks.
}
