% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{do_alignPEAKS}
\alias{do_alignPEAKS}
\title{Align dataset's and template peaks using dot-product estimation}
\usage{
do_alignPEAKS(ds, tmp, ds_var_name, tmp_var_name, mz_err, rt_err, bins,
  ncores, cutoff, anno = FALSE)
}
\arguments{
\item{ds}{\code{data.frame} containing peaks which have to be compared and aligned with the template.}

\item{tmp}{\code{data.frame} representing a template with which the dataset is compared to.}

\item{ds_var_name}{\code{character} indicating the column name for peak grouping information in the dataset.
Default is set to 'peakgr', which is the output of \code{groupPEAKS} method.}

\item{tmp_var_name}{\code{character} indicating the column name for peak grouping information in the template.
If function is used within \code{alignPEAKS}, then 'peakgr' should be selected.
If function is used within \code{annotatePEAKS}, then 'chemid' should be selected.}

\item{mz_err}{\code{numeric} specifying the window for peak matching in the MZ dimension.}

\item{rt_err}{\code{numeric} specifying the window for peak matching in the RT dimension.}

\item{bins}{\code{numeric} defying step size used in peak-group spectra binning and vector generation. Step size represents MZ dimension.}

\item{ncores}{\code{numeric} for number of parallel workers to be used.}

\item{cutoff}{\code{numeric} for spectra similarity score threshold, set to 0.5 by default.}

\item{anno}{\code{logical} whether function is used to align samples or to annotate final features with a database. Set to FALSE by default.}
}
\value{
Function returns a list of lenght of the number of peak-groups in the dataset.
Each list entry specifies the template's peak-group to which dataset's peak-group was aligned to, as well as matching peaks and obtained cosine.
}
\description{
Internal function used within method \code{\link{alignPEAKS}} and function \code{\link{annotateDS}}.
Function aligns all dataset's peaks to the template using dot-product estimation.
}
\seealso{
For details on cosine estimation, refer to \code{\link{alignPEAKS}} method.
}
