% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-validPEAKS_helpers.R
\name{exportPEAK}
\alias{exportPEAK}
\title{Export centWave measures for a peak-of-interest in every sample in which it was detected.}
\usage{
exportPEAK(peakid, peaks_vals_samples)
}
\arguments{
\item{peakid}{\code{numeric} corresponding to peak ID in the final template.}

\item{peaks_vals_samples}{\code{list} with centWave measures for every sample, with length equal to sample number.}
}
\value{
Function return a \code{data.frame} with centWave measures for a peak-of-interest, for every sample in the provided peaks_vals_samples list.
}
\description{
Export centWave measures for a peak-of-interest in every sample in which it was detected.
}
