% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-validPEAKS_helpers.R
\name{exportSAMPLE}
\alias{exportSAMPLE}
\title{Extract centWave values for peaks in the final template for a sample-of-interest.}
\usage{
exportSAMPLE(sdata, final_tmp)
}
\arguments{
\item{sdata}{\code{data.frame} with centWave measures for the sample-of-interest.}

\item{final_tmp}{\code{data.frame} with columns "peakid" and "pcs" indicating to which Pseudo Chemical Spectra each peak (peakid) belongs to.
Contains only the peaks that were returned by peak validation algorithm.}
}
\value{
Function returns a \code{data.frame} with all centWave values for all peaks from the final template that were detected in the sample-of-interest.
}
\description{
Function extracts centWave peak-picking values for a single sample.
Only peaks in the final template are returned.
}
\seealso{
\code{\link{validPEAKGR}}, \code{\link{validPEAKS}}
}
