% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-validPEAKS_helpers.R
\name{extractCOMMUNITIES}
\alias{extractCOMMUNITIES}
\title{Extract communities, or, Pseudo Chemical Spectra, with > 1 peak}
\usage{
extractCOMMUNITIES(peakgrs_split)
}
\arguments{
\item{peakgrs_split}{\code{list} for every community into which peak-group was split to by the Cluster Label Propagation algorithm.}
}
\value{
Function returns a \code{data.frame} with peakids and corresponding Pseudo Chemical Spectra (PCS).
}
\description{
Function extracts communities, or, Pseudo Chemical Spectra, with > 1 peak, and returns final peak table.
}
\seealso{
\code{\link{validPEAKGR}}, \code{\link{validPEAKS}}
}
