% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-fillPEAKS_helpers.R
\name{extractINT}
\alias{extractINT}
\title{Extract raw signal values for defined mz and rt region.}
\usage{
extractINT(raw, rt, mz)
}
\arguments{
\item{raw}{\code{OnDiskMSnExp} class object.}

\item{rt}{\code{data.frame} with rtmin and rtmax values (columns) for peaks (rows) that have to be re-integrated in the sample-of-interest.}

\item{mz}{\code{data.frame} with mzmin and mzmax values (columns) for peaks (rows) that have to be re-integrated in the sample-of-interest.}
}
\value{
Function returns a \code{list} of data.frames with extracted raw LC-MS signal values.
}
\description{
Helper function used by \code{\link{fillSAMPLE}} function to extract spectrum intensity values for a defined mz and rt window.
}
