% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-fillPEAKS_helpers.R
\name{extractMISS}
\alias{extractMISS}
\title{Extract samples with missed peaks, for which intensity have to be re-integrated.}
\usage{
extractMISS(s, values, peaks_mod, valid)
}
\arguments{
\item{s}{\code{numeric} corresponding to sample number in the study.}

\item{values}{\code{list} with centWave values for each sample.}

\item{peaks_mod}{\code{list} with real and modelled centWave values for each peak.}

\item{valid}{\code{data.frame} with validated aligned peaks.}
}
\value{
Function returns a list with a data.frame for each sample in the study.
Each data.frame contains predicted mz and rt regions only for those peaks that were missed in the corresponding sample.
}
\description{
Function builds a list with mz and rt regions for intensity re-integration.
This is a data-reduction step, helping to reduce the size of the variables being imputed to \code{\link{fillSAMPLE}} function for intensity re-integration.
}
