% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowDB.R
\docType{methods}
\name{filepath,massFlowDB-method}
\alias{filepath,massFlowDB-method}
\title{Get the the absolute path to the chemical database file}
\usage{
\S4method{filepath}{massFlowDB}(object)
}
\arguments{
\item{object}{\code{massFlowDB} class object}
}
\description{
Get the the absolute path to the chemical database file
}
