% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\name{filepath,massFlowTemplate-method}
\alias{filepath,massFlowTemplate-method}
\alias{filepath}
\title{Obtain the absolute path to metadata file of the experiment}
\usage{
\S4method{filepath}{massFlowTemplate}(object)
}
\arguments{
\item{object}{\code{massFlowTemplate} class object}
}
\description{
Obtain the absolute path to metadata file of the experiment.
}
\seealso{
\code{\link{massFlowTemplate-class}}
}
