% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\docType{methods}
\name{fillPEAKS,massFlowTemplate-method}
\alias{fillPEAKS,massFlowTemplate-method}
\alias{fillPEAKS}
\title{Fill peaks}
\usage{
\S4method{fillPEAKS}{massFlowTemplate}(object, out_dir = NULL,
  ncores = 1)
}
\arguments{
\item{object}{\code{massFlowTemplate} class object.}

\item{out_dir}{\code{character} specifying desired directory for output.}

\item{ncores}{\code{numeric} defining number of cores to use for parallelisation. Default set to 1 for serial implementation.}

\item{min_samples}{\code{numeric} specifying the minimum percentage of samples in which peak has to be detected in order to be considered (default set to 10 percent).}

\item{cor_thr}{\code{numeric} defining Pearson correlation coefficient threshold for inter-sample correlation between peaks (default set to 0.75).}
}
\value{
Method returns peak table with filled intensities for missed peaks.
}
\description{
Fill peaks
}
\seealso{
\code{\link{loadALIGNED}}, \code{\link{alignPEAKS}}, \code{\link{validPEAKS}}, \code{\link{massFlowTemplate-class}}
}
