% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-fillPEAKS_helpers.R
\name{fillSAMPLE}
\alias{fillSAMPLE}
\title{Fill intensity values for peaks missed in the sample-of-interest.}
\usage{
fillSAMPLE(s, sname, sdata, values)
}
\arguments{
\item{s}{\code{numeric} corresponding to the number of the sample-of-interest in the study.}

\item{sname}{\code{character} with full path to the raw LC-MS file for the sample-of-interest.}

\item{sdata}{\code{data.frame} with all centWave values for the sample-of-interest.}

\item{values}{\code{data.frame} with predicted mz and rt regions only for those peaks that were missed in the corresponding sample.}
}
\value{
Function returns updated \code{data.frame} with the centWave values for the sample-of-interest.
}
\description{
Fill intensity values for peaks missed in the sample-of-interest.
}
