% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowAnno.R
\docType{methods}
\name{findANNOchemid,massFlowAnno-method}
\alias{findANNOchemid,massFlowAnno-method}
\alias{findANNOchemid}
\title{Find annotations for selected chemical compound id}
\usage{
\S4method{findANNOchemid}{massFlowAnno}(object, chemid = NULL,
  cutoff = 0)
}
\arguments{
\item{object}{\code{massFlowAnno} class object.}

\item{chemid}{\code{numeric} specifying one chemical compound id to look at.}

\item{cutoff}{\code{numeric} specifying cosine threshold, annotations below are not reported. Default set to 0.}
}
\value{
Method returns a \code{data.frame} with all valid pseudo chemical spectra annotated to the selected chemical reference database compound.
}
\description{
Method returns all valid annotations for the selected chemical compound.
}
