% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowAnno.R
\docType{methods}
\name{findANNOpcs,massFlowAnno-method}
\alias{findANNOpcs,massFlowAnno-method}
\alias{findANNOpcs}
\title{Find annotations for selected pseudo chemical spectra}
\usage{
\S4method{findANNOpcs}{massFlowAnno}(object, pcs = NULL, cutoff = 0)
}
\arguments{
\item{object}{\code{massFlowAnno} class object.}

\item{pcs}{\code{numeric} specifying one pseudo chemical spectra to look at.}

\item{cutoff}{\code{numeric} specifying cosine threshold, annotations below are not reported. Default set to 0.}
}
\value{
Method returns a \code{data.frame} with all valid chemical reference database compounds annotated to the selected pseudo chemical spectra.
}
\description{
Method returns all valid annotations for the selected pseudo chemical spectra.
}
