% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-getCLUSTS.R
\name{getCLUSTS}
\alias{getCLUSTS}
\title{Obtain clusters of components for each datafile}
\usage{
getCLUSTS(files, out_dir, write = T, workers = 2)
}
\arguments{
\item{workers}{}
}
\description{
Obtain clusters of components for each datafile
}
