% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-getCOMPS.R
\name{getCOMPS}
\alias{getCOMPS}
\title{Obtain components for all datafiles}
\usage{
getCOMPS(files, out_dir, cwt, match = 1, pearson = TRUE, thr = 0.95,
  plot = FALSE, clean = TRUE, bpparam)
}
\arguments{
\item{workers}{}
}
