% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCORmat}
\alias{getCORmat}
\title{Build a correlation matrix between peaks-of-interest}
\usage{
getCORmat(ind)
}
\arguments{
\item{ind}{\code{numeric} with indeces of peaks-of-interest.}
}
\value{
Function returns a \code{data.frame} with unique peak-pairs.
}
\description{
Function builds an matrix for pairs between provided peaks-of-interest.
Each row is a unique peak-pair with columns 'from' and 'to' specifying peak indeces.
}
