% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{getCOSmat}
\alias{getCOSmat}
\title{Estimate cosine angles between the vectors representing dataset and template peak-groups}
\usage{
getCOSmat(ds_bin, ds_var, tmp_bin, tmp_var, mz_err, rt_err, bins = 0.01)
}
\arguments{
\item{ds_bin}{\code{data.frame} for the single \emph{rt} region of the dataset.}

\item{ds_var}{\code{character} indicating the column name for peak grouping information in the dataset.}

\item{tmp_bin}{\code{data.frame} for the single \emph{rt} region of the template}

\item{tmp_var}{\code{character} indicating the column name for peak grouping information in the template.}

\item{mz_err}{\code{numeric} specifying the window for peak matching in the MZ dimension.}

\item{rt_err}{\code{numeric} specifying the window for peak matching in the RT dimension.}

\item{bins}{\code{numeric} defying step size used in peak-group spectra binning and vector generation.#'}
}
\value{
Function returns a list with two entries:
\itemize{
 \item \code{matrix} with obtained cosines between all dataset and template peak-groups.
 \item \code{list} with matching template peaks for every peak-group in the dataset.
 }
}
\description{
Function estimates cosine angles between matching dataset and template peak-groups.
Function is applied to a single \emph{rt} region(bin) and is used in parallel on multiple \emph{rt} regions.
Function is applied to \emph{rt} region within \code{\link{do_alignPEAKS}} function.
}
