% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-validPEAKS_helpers.R
\name{getINT}
\alias{getINT}
\title{Get intensity of a peak-of-interest in a single sample.}
\usage{
getINT(s, peak)
}
\arguments{
\item{s}{\code{numeric} indicating sample's run order.}

\item{peak}{\code{data.frame} with intensities of the peak-of-interest in all samples.}
}
\value{
Function returns an \code{integer} with peak intensities.
}
\description{
Get intensity of a peak-of-interest in a single sample.
}
