% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{getMATCHES}
\alias{getMATCHES}
\title{Get template peaks that match the target peak by \emph{m/z} and \emph{rt} values}
\usage{
getMATCHES(target_peak, tmp, tmp_var, target_var)
}
\arguments{
\item{target_peak}{\code{matrix} with target peak's \emph{m/z} and \emph{rt} lower and higher values.}

\item{tmp}{\code{data.frame} representing a template in which matching peaks are looked for.}

\item{tmp_var}{\code{character} indicating the column name for peak grouping information in the template.}

\item{target_var}{\code{character} indicating the column name for peak grouping information in the target peak's \code{matrix}.}
}
\value{
Function returns a \code{data.frame} with matching template's peaks.
}
\description{
Function find all matching peaks in a template.
Broad \emph{m/z} and \emph{rt} values are used in the search.
Function is applied to every peak in a peak-group of interest.
}
