% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-validPEAKS_helpers.R
\name{getPEAKIDS}
\alias{getPEAKIDS}
\title{Extract peakids from every community/Pseudo Chemical Spectra of a peak-group-of-interest.}
\usage{
getPEAKIDS(pkg)
}
\arguments{
\item{pkg}{\code{list} with communities into which each peak-group was split to.
List length is equal to number of peak-groups.}
}
\value{

}
\description{
Function is used after \code{\link{validPEAKGR}} has been applied to identify communities of peaks using correlation network analysis.
Function returns peakids of every community into which a peak-group-of-interest was split to.
}
\seealso{
\code{\link{validPEAKGR}}, \code{\link{validPEAKS}}.
}
