% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-validPEAKS_helpers.R
\name{getPEAKmedians}
\alias{getPEAKmedians}
\title{Get median centWave measures for a peak-of-interest across all samples in which it was detected.}
\usage{
getPEAKmedians(peak_n, values = c("mz", "mzmin", "mzmax", "rt", "rtmin",
  "rtmax"))
}
\arguments{
\item{peak_n}{\code{data.frame} with centWave measures for a peak-of-interest in every sample}
}
\value{
Function return a \code{data.frame} with median centWave measures for a peak-of-interest across all samples.
}
\description{
Get median centWave measures for a peak-of-interest across all samples in which it was detected.
}
