% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-joinPEAKS_helpers.R
\name{getRSE}
\alias{getRSE}
\title{Estimate RSE for a polynomial model with given peaks combination}
\usage{
getRSE(combs, mat_dt, t_peakid)
}
\arguments{
\item{combs}{\code{integer} with matching peaks peakids.}

\item{mat_dt}{\code{data.frame} with all peaks that belong to a given sub-peak.}

\item{t_peakid}{\code{integer} with the current target peak's peakid.}
}
\value{
Function returns RSE value for the model built with the given peaks combination.
}
\description{
Function build a polynomial model using the RT values for the target and its matching peaks combination. The residual standard error is returned.
}
