% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{getRTbins}
\alias{getRTbins}
\title{Split a dataset-of-interest and template into \emph{rt} regions (bins) for parallelisation}
\usage{
getRTbins(ds, tmp, ds_var_name, tmp_var_name, mz_err, rt_err, ncores)
}
\arguments{
\item{ds}{\code{data.frame} containing peaks which have to be compared and aligned with the template.}

\item{tmp}{\code{data.frame} representing a template with which the dataset is compared to.}

\item{ds_var_name}{\code{character} indicating the column name for peak grouping information in the dataset.
Default is set to 'peakgr', which is the output of \code{groupPEAKS} method.}

\item{tmp_var_name}{\code{character} indicating the column name for peak grouping information in the template.
If function is used within \code{alignPEAKS}, then 'peakgr' should be selected.
If function is used within \code{annotateDS}, then 'chemid' should be selected.}

\item{mz_err}{\code{numeric} specifying the window for peak matching in the MZ dimension.}

\item{rt_err}{\code{numeric} specifying the window for peak matching in the RT dimension.}

\item{ncores}{\code{numeric} for number of parallel workers to be used.}
}
\value{
Function returns a list with dataset and template tables split into \emph{rt} regions.
}
\description{
Function splits a dataset and a template into \emph{rt} regions, each of which is analysed by an independent parallel worker.
}
\details{
Function firstly splits the dataset into as many sub-tables as there are parallel workers (could be 1, if serial implementation is selected).
Then template is split into the same number of sub-tables using \emph{rt} regions in the corresponding dataset's sub-tables.
}
