% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupCOMPS.R
\name{groupCOMPS}
\alias{groupCOMPS}
\title{Group components across datafiles.}
\usage{
groupCOMPS(files, mz_err = 0.01, rt_err = 0.2, bins = 0.01)
}
\arguments{
\item{files}{A \code{character} with paths to peak tables with components and clusters. Created by \emph{buildCOMPS} function.}

\item{mz_err}{A \code{numeric} specifying the window for peak matching in the MZ dimension. Default set to 0.01.}

\item{rt_err}{A \code{numeric} specifying the window for peak matching in the RT dimension. Default set to 0.2 (min).}

\item{bins}{A \code{numeric} defying step size used in component's spectra binning and vector generation. Step size represents MZ dimension (default set to 0.01).}
}
\description{
Group components across datafiles.
}
