% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupPEAKS.R
\name{groupPEAKS}
\alias{groupPEAKS}
\title{Get peak-groups representing different chemical spectra in each LC-MS datafile}
\usage{
groupPEAKS(files, out_dir, cwt, match = 1, pearson = TRUE,
  thr = 0.95, plot = FALSE, clean = TRUE, bpparam)
}
\arguments{
\item{files}{\code{character} with paths to mzML file(s) to be processed.}

\item{out_dir}{\code{character} specifying desired directory for output.}

\item{cwt}{\code{CentWaveParam} class object with parameters for centWave-based peak-picking.}

\item{match}{\code{numeric} defining the number of scans for co-eluting peaks extraction.}

\item{pearson}{A \code{logical} whether Pearson Correlation should be used. For \code{pearson = FALSE}, Spearman correlation method will be used.}

\item{thr}{A \code{numeric} defining correlation coefficient threshold, above which peak pairs will be considered as correlated.}

\item{plot}{A \code{logical}. For \code{plot = TRUE}, a network graph for each peak in the table will be saved as a png file in the out_dir directory.}

\item{clean}{A \code{logical} whether one-peak peak-groups should be removed (default is TRUE).}

\item{bpparam}{A \code{BiocParallel} parameter object to control how and if parallel processing should be performed.
Such object can be created by the \emph{SerialParam}, \emph{MulticoreParam} or \emph{SnowParam} functions from the \emph{BiocParallel} package.}
}
\value{
For each LC-MS file, function writes a table with picked-peaks and their peak-groups into separate files, in the defined \code{out_dir} directory.
}
\description{
Function performs peak-picking and peak-grouping for each LC-MS datafile independently.
}
\details{
Function performs peak-picking using the \emph{centWave} algorithm from package \emph{xcms}.
Picked-peaks are then grouped into chemical spectra. For each peak in the sample:
\itemize{
\item Co-eluting peaks are found.
\item EIC correlation between all co-eluting peaks is performed.
\item Network of peaks with high EIC correlation (with coefficients above the selected threshold) is built.
}
Co-eluting peaks within a network of high EIC correlation originate from the same chemical compound and therefore form a chemical spectrum.
}
