% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\name{joinPEAKS,massFlowTemplate-method}
\alias{joinPEAKS,massFlowTemplate-method}
\alias{joinPEAKS}
\title{joinPEAKS}
\usage{
\S4method{joinPEAKS}{massFlowTemplate}(object, out_dir = NULL, mz_err = 0.001, rt_err = 1, ncores = 2)
}
\value{
Method returns \code{massFlowTemplate} class object with updated template and samples data.
}
\description{
Development mode, unfinished.

Method takes validated pseudo chemical spectra
}
