% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{loadALIGNED}
\alias{loadALIGNED}
\title{Build sample alignment and annotation template using already aligned samples}
\usage{
loadALIGNED(
  file = NULL,
  template = NULL,
  mz_err = 0.01,
  rt_err = 2,
  bins = 0.05,
  realtime = FALSE
)
}
\arguments{
\item{file}{A \code{character} with path to the csv file, specifying samples filenames and their acquisition order.}

\item{template}{A \code{character} with path to the csv file with the latest template obtained by \code{alignPEAKS} function.}

\item{mz_err}{A \code{numeric} specifying the window for peak matching in the MZ dimension. Default set to 0.01.}

\item{rt_err}{A \code{numeric} specifying the window for peak matching in the RT dimension. Default set to 2 (sec).}

\item{bins}{A \code{numeric} defying step size used in component's spectra binning and vector generation. Step size represents MZ dimension (default set to 0.05).}

\item{realtime}{\code{logical} whether real-time implementation is required. If set to TRUE, alignment will wait for intermediate peakgroups files to be written.}
}
\value{
A \code{massFlowTemplate} class object.
}
\description{
Function handles the construction of \code{massFlowTemplate} class object from already aligned samples.
Function enables user to continue an interrupted peak alignment process, facilitated by \code{alignPEAKS} function.
}
\details{
Arguments are identical to the ones used by \code{\link{buildTMP}} constructor function.
}
\seealso{
\code{\link{massFlowTemplate}} class.
}
