% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/functions-validmassFlowAnno.R,
%   R/methods-massFlowAnno.R
\docType{class}
\name{massFlowAnno-class}
\alias{massFlowAnno-class}
\alias{massFlowAnno}
\alias{validmassFlowAnno}
\alias{show,massFlowAnno-method}
\title{massFlowAnno, a class for peak table annotation with a chemical standards database.}
\usage{
validmassFlowAnno(object)

\S4method{show}{massFlowAnno}(object)
}
\arguments{
\item{object}{\code{massFlowAnno} class object}
}
\description{
This class enables annotation of the final peak table obtained with massFlowR pipeline.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{\code{character} specifying the absolute path to \code{csv} file with peak intensity table.}

\item{\code{samples}}{\code{data.frame} storing study sample names and their acquisition order.}

\item{\code{data}}{\code{data.frame} with peak intensity values for each sample.}

\item{\code{ds}}{\code{data.frame} with pseudo-chemical-spectra intensity values of samples with the highest intensity for that PCS.}

\item{\code{db}}{\code{data.frame} storing chemical standards database table, obtained with \code{\link{buildDB}}.}

\item{\code{anno}}{\code{data.frame} storing peak table with database annotations (if any).}

\item{\code{mat}}{\code{matrix} with peak annotation results, rows are chemid (database chemicals) and columns are PCS.}

\item{\code{params}}{\code{list} containg database matching parameters.}
}}

\seealso{
\code{\link{buildDB}}, \code{\link{annotateDS}}
}
