% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods-massFlowDB.R
\docType{class}
\name{massFlowDB-class}
\alias{massFlowDB-class}
\alias{massFlowDB}
\alias{show,massFlowDB-method}
\alias{filepath,massFlowDB-method}
\alias{chemicals,massFlowDB-method}
\title{Chemical reference database template}
\usage{
\S4method{show}{massFlowDB}(object)

\S4method{filepath}{massFlowDB}(object)

\S4method{chemicals}{massFlowDB}(object)
}
\arguments{
\item{object}{\code{massFlowDB} class object}
}
\description{
\code{massFlowDB} class encapsulates information provided in the chemical reference database.

Function returns the absolute path to the \code{csv} file used when building the \code{massFlowDB} object.
File contains chemical reference database.

Function returns names of the chemicals in the database.
}
\details{
Chemical reference database file lists peaks \code{centWave}-detected in the reference LC-MS datafiles.
The database file is a single csv file, which must contain columns: "peakid", "mz", "rt", "into", "peakgr", "chemid", "dbid", "dbname".

To initialise \code{massFlowDB} object, a constructor function \code{\link{buildDB}} mus be used.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{\code{character} specifying path to database csv file.}

\item{\code{db}}{\code{data.frame} containing database template.}
}}

\seealso{
\code{\link{buildDB}}
}
