% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{massFlowDB}
\alias{massFlowDB}
\title{Build a chemical database template.}
\usage{
massFlowDB(file = NULL)
}
\arguments{
\item{file}{A \code{character} with path to the the database template file (csv).}
}
\value{
A \code{massFlowDB} class object.
}
\description{
Functions prepares a \code{massFlowDB} class object, which will be used for chromatograpic peaks annotations in study samples.
}
\examples{
## Build template using a sample, provided with the package
data_dir <- system.file("testdata", package = "massFlowR")
db_fname <- file.path(data_dir, "DBtemplate.csv")
db <- massFlowDB(file = db_fname)

}
