% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massFlowR.R
\docType{package}
\name{massFlowR-package}
\alias{massFlowR}
\alias{massFlowR-package}
\title{massFlowR: a package for pre-processing of high-resolution, untargeted, centroid LC-MS data}
\description{
massFlowR annotates and aligns structurally-related spectral peaks across LC-MS experiment samples.
Its pipeline consists of three stages:
}
\details{
\itemize{
\item Individual samples processing
\item Multiple sample alignment
\item Alignment validation and peak filling
}
}
