% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-fillPEAKS_helpers.R
\name{modelPEAKS}
\alias{modelPEAKS}
\title{Function models intensity integration values for missing peaks.}
\usage{
modelPEAKS(p, vars, object)
}
\arguments{
\item{p}{\code{numeric} indicating peak IDs.}

\item{vars}{\code{character} with centWave values to be modelled, e.g. c("mz", "mzmin", "mzmax", "rt", "rtmin", "rtmax").}

\item{object}{\code{massFlowTemplate} class object.}
}
\value{
Function returns a list of data.frames - a data.frame for each peak supplied with the argument p.
}
\description{
Function models intensity integration values for each of the missing peaks and for each of the samples in which it was missed.
Modelling relies on a smooth spline fit to the values of the corresponding detected peaks.
}
