% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{normSPEC}
\alias{normSPEC}
\title{Normalise a vector representing peaks' intensity values to the total magnitude of the vector}
\usage{
normSPEC(into)
}
\arguments{
\item{into}{\code{numeric} vector with intensity values.}
}
\value{
Function returns a \code{numeric} vector with normalised intensity values.
}
\description{
Function normalises peaks' intensity values to the total magnitude of the vector,
as decribed in Lam, Henry, et al. 2007
}
