% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{orderBYrt}
\alias{orderBYrt}
\title{Order a peak table by variables' \emph{rt}}
\usage{
orderBYrt(dt, var_name)
}
\arguments{
\item{dt}{\code{data.frame} of the peak table of interest.}

\item{var_name}{\code{character} specifying column name for variables which should be ordered by their \emph{rt}.
Depending on the step in the pipeline, variable name will be either 'peakgr', 'chemid' or 'pcs'.}
}
\value{
Function returns ordered \code{data.frame}.
}
\description{
Order a peak table by variables' \emph{rt}
}
