% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\name{peaksVALIDATED,massFlowTemplate-method}
\alias{peaksVALIDATED,massFlowTemplate-method}
\alias{peaksVALIDATED}
\title{Check if massFlowTemplate object was validated}
\usage{
\S4method{peaksVALIDATED}{massFlowTemplate}(object)
}
\arguments{
\item{object}{\code{massFlowTemplate} class object}
}
\description{
Method returns TRUE if \code{massFlowTemplate} object has been validated via \code{\link{validPEAKS}} method.
}
\seealso{
\code{\link{massFlowTemplate-class}}
}
