% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-pickPEAKS.R
\name{pickPEAKS}
\alias{pickPEAKS}
\title{Title}
\usage{
pickPEAKS(raw, ppm, snthresh, noise, prefilter, peakwidth, integrate, verbose,
  fitGauss, clean = TRUE)
}
\arguments{
\item{clean}{}
}
\description{
Title
}
