% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowAnno.R
\docType{methods}
\name{plotCHEMID,massFlowAnno-method}
\alias{plotCHEMID,massFlowAnno-method}
\title{Plot selected database chemical compound.}
\usage{
\S4method{plotCHEMID}{massFlowAnno}(object, chemid = NULL,
  cutoff = 0.5, out_dir = NULL)
}
\arguments{
\item{object}{\code{massFlowAnno} class object.}

\item{chemid}{\code{numeric} specifying chemical compound id to look at.}

\item{cutoff}{\code{numeric} specifying spectral similarity score value.}

\item{out_dir}{\code{character} specifying desired directory for generated figure. Default set to NULL and figure is plotted to graphical device.}
}
\value{
Method returns a plot with the spectra of selected chemical compound and top five annotated pseudo chemical spectra from the dataset.
}
\description{
Method returns a plot with the selected chemical compound's spectra and top five annotated pseudo chemical spectra in the dataset.
If out_dir is provided, generated plot is written as a png file. Otherwise, spectra is plotted on current graphical device.
\code{massFlowAnno} class object must be annotated using \code{\link{annotateDS}} method first.
}
