% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowAnno.R
\docType{methods}
\name{plotPCS,massFlowAnno-method}
\alias{plotPCS,massFlowAnno-method}
\alias{plotPCS}
\title{Plot selected pseudo chemical spectra}
\usage{
\S4method{plotPCS}{massFlowAnno}(object, pcs = NULL, anno = FALSE,
  cutoff = 0, out_dir = NULL)
}
\arguments{
\item{object}{\code{massFlowAnno} class object.}

\item{pcs}{\code{numeric} specifying one or more pseudo chemical spectra to look at.}

\item{anno}{\code{logical} whether spectra of annotated database compounds should be plotted.}

\item{cutoff}{\code{numeric} specifying spectral similarity score value.}

\item{out_dir}{\code{character} specifying desired directory for generated figure. Default set to NULL and figure is plotted to graphical device.}
}
\value{
Method returns a plot with the spectra of single pseudo chemical spectra.
}
\description{
Method returns a plot with the selected single pseudo chemical spectra in the sample in which it was most intense.
If anno set to TRUE, spectra of top five annotated database compounds are also plotted.
If out_dir is provided, generated plot is written as a png file. Otherwise, spectra is plotted on current graphical device.
\code{massFlowAnno} class object must be annotated using \code{\link{annotateDS}} method first.
}
