% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{rankCOS}
\alias{rankCOS}
\title{Rank cosines, assigning 1 to the highest cosine and 0s to cosine of 0.}
\usage{
rankCOS(x)
}
\arguments{
\item{x}{\code{matrix} with cosine values for a single peak-group of interest and all peak-groups in table being matched to.}
}
\value{
Function returns \code{numeric} vector with ranks for each peak-group.
Of the same lenght as original x \code{matrix}.
}
\description{
Rank cosines, assigning 1 to the highest cosine and 0s to cosine of 0.
}
