% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbindCLEAN}
\alias{rbindCLEAN}
\title{Bind a list of data frames}
\usage{
rbindCLEAN(...)
}
\arguments{
\item{...}{}
}
\value{
Function returns a data frame.
}
\description{
Function binds a list of data frames into a single data frame and removes inherited row names.
}
