% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readDATA}
\alias{readDATA}
\title{Read raw LC-MS data into memory}
\usage{
readDATA(f)
}
\arguments{
\item{f}{\code{character} specifying absolute path to a single raw LC-MS file.}
}
\value{
Function returns \code{OnDiskMSnExp} class object.
}
\description{
Function reads raw LC-MS datafile into memory using \code{MSnbase} functionality.
}
