% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaleEDGES}
\alias{scaleEDGES}
\title{Scale correlation coefficients}
\usage{
scaleEDGES(x, from = 0.01, to = 10)
}
\arguments{
\item{x}{\code{numeric} specifying correlation coefficients of a graph.}

\item{from}{\code{numeric} specifying the lowest value to which graph weights should be scaled to.}

\item{to}{\code{numeric} specifying the highest value to which graph weights should be scaled to.}
}
\value{
Functions returns scaled graph's weights in \code{numeric} format.
}
\description{
Function scales correlation coefficients to make a clear Fruchterman-Reingold graph.
}
