% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-fillPEAKS_helpers.R
\name{smoothVALUE}
\alias{smoothVALUE}
\title{Helper function to model intensity integration values for missing peaks.}
\usage{
smoothVALUE(var, x, peak, lambda = 10)
}
\arguments{
\item{var}{\code{character} with a single centWave value to be modelled, e.g. "mz"}

\item{x}{\code{numeric} representing indeces of peaks for the supplied 'peak' data.frame.}

\item{peak}{\code{data.frame} with all centWave values (columns) for the peak-of-interest for all samples (rows). For samples in which peak as not detected, NAs must be listed.}

\item{lambda}{\code{numeric} representing smoothing parameter lambda. Set to 10 by default, which corresponds to number of study samoples in between two quality control samples in standard experimental protocols.}
}
\value{
Function returns centWave values for the peak-of-interest, which includes predicted values for samples in which it was not originally detected.
}
\description{
Helper function is used by the \code{\link{modelPEAKS}} function to model intensity integration values for missing peaks.
Function fit a smoothing spline to centWave values reported for the peak-of-interest in all samples in the study.
For samples in which peak was not detected, centWave values are predicted.
}
