% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-validPEAKS_helpers.R
\name{validPEAKGR}
\alias{validPEAKGR}
\title{Validate a peak-group-of-interest by performing a cross-sample Pearson correlation.}
\usage{
validPEAKGR(pkg, peakgrs_ints, out_dir, cor_thr, min_samples_n,
  save_plot = FALSE)
}
\arguments{
\item{pkg}{\code{numeric} specifying a peak-group-of-interest.}

\item{peakgrs_ints}{\code{list} with centWave values, listed by peak-groups.}

\item{out_dir}{\code{character} specifying desired directory for output, if plotting is selected.}

\item{cor_thr}{\code{numeric} defining Pearson correlation coefficient threshold for inter-sample correlation between peaks.}

\item{min_samples_n}{\code{numeric} specifying the minimum number of samples in which peak has to be detected in order to be considered.}

\item{save_plot}{\code{logical} whether network plot should be saved as a pdf file in the out_dir, default set to FALSE.}
}
\value{
Function returns a \code{list} for every community into which peak-group was split to by the Cluster Label Propagation algorithm.
Each element of the returned list contains centWave values for every community peak in the samples in which it was detected.
}
\description{
Functions correlates all peaks of a peak-groups across all samples in which they were detected.
Obtained Pearson correlation coefficients are used to build a peak-peak network with pairs above selected correlation threshold.
\emph{igraph} Cluster Label Propagation algorithm is used to detect communities of peaks with the network.
These communities are returned as independent peak-goups, or, Pseudo Chemical Spectra.
}
