% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-addDOI.R
\name{addDOI}
\alias{addDOI}
\title{Align and add peaks from a sample to the template}
\usage{
addDOI(tmp, tmp_fname, doi_fname, mz_err, rt_err, bins, write_int)
}
\arguments{
\item{tmp}{\code{data.frame} with latest template version.}

\item{tmp_fname}{\code{character} specifying absolute path where updated template will be written to.}

\item{doi_fname}{\code{character} specifying absolute path of the datafile-of-interest csv file.}

\item{mz_err}{\code{numeric} specifying the window for peak matching in the MZ dimension. Default set to 0.01.}

\item{rt_err}{\code{numeric} specifying the window for peak matching in the RT dimension. Default set to 2 (sec).}

\item{bins}{\code{numeric} defying step size used in component's spectra binning and vector generation. Step size represents MZ dimension (default set to 0.1).}

\item{write_int}{\code{logical} whether resulting template should be written to a separate file. Default set to FALSE, so that template is over-write to the same filename.}
}
\value{
Function returns updated TMP and DOI.
Both TMP and DOI are written to a csv file in the same directory as the original peak table for the sample.
}
\description{
Function is used by method \code{\link{alignPEAKS}}.
Function aligns all peaks in a datafile-of-interest (DOI) with the template (TMP).
Many of the required parameters are identical to the constructor function \code{\link{buildTMP}}.
}
\seealso{
\code{\link{alignPEAKS}} method.
}
