% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{assignCOS}
\alias{assignCOS}
\title{Assign dataset's peak-groups to template peak-groups based on estimated cosine angles}
\usage{
assignCOS(cos)
}
\arguments{
\item{cos}{\code{matrix} with cosine angles.
Rows correspond to template variables (i.e. peak-group).
Columns correspond to dataset-of-interest variables (i.e. peak-group).}
}
\value{
Function returns a \code{matrix} of the same dimension.
Peak-group pairs which should be aligned are marked with TRUE in the output matrix.
}
\description{
Assign dataset's peak-groups to template peak-groups based on estimated cosine angles
}
