% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{assignCOS}
\alias{assignCOS}
\title{Assign dataset's peak-groups to template peak-groups based on estimated cosine angles}
\usage{
assignCOS(cos, cutoff = 0)
}
\arguments{
\item{cos}{\code{matrix} with cosine angles.
Rows correspond to template variables (i.e. peak-group).
Columns correspond to dataset-of-interest variables (i.e. peak-group).}

\item{cutoff}{\code{numeric} for spectra similarity score threshold, set to 0 by default.}
}
\value{
Function returns a \code{matrix} of the same dimensions.
Peak-groups which should be grouped/aligned into one are marked with TRUE in the output matrix.
}
\description{
Maximise peak-groups assignment by selecting highest-scoring match for each peak-group in the template.
}
\details{
Function is used for peak-group alignment by \code{\link{alignPEAKS}} method and for automatic annotation on \code{\link{massFlowAnno-class}} object.
}
