% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-buildCOMPS.R
\name{buildCOMPS}
\alias{buildCOMPS}
\title{Component building using correlation estimation between co-eluting peaks}
\usage{
buildCOMPS(Pearson, match = 1, pks, thr = 0.95, plot = FALSE,
  clean = TRUE)
}
\arguments{
\item{Pearson}{logical whether correlation between co-eluting peaks should be estimated using Pearson Correlation. If FALSE, Spearman will be used.}

\item{match}{numeric defining the number of scans for co-eluting peaks extraction.}

\item{pks}{a data.frame object, provided by pickPEAKS() function.}

\item{thr}{numeric defining correlation coefficient threshold, above which peak pairs will be considered as correlated.}

\item{plot}{logical. If TRUE, a png plot for each peak and all of its co-eluting peaks will be saved.}
}
\seealso{
\code{\link{pickPEAKS}}
}
