% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-buildDB.R
\name{buildDB}
\alias{buildDB}
\title{Build a chemical reference database table using rda files}
\usage{
buildDB(rda_dir = NULL, out_dir = NULL, peakgr_thr = 0,
  compound_thr = 0)
}
\arguments{
\item{rda_dir}{\code{character} with absolute path to the directory with rda files.}

\item{out_dir}{\code{character} with absolute path to the directory where database file should be written.}

\item{peakgr_thr}{\code{numeric} specifying intensity threshold (100\% BPI) for peaks in a chemical spectrum. If set to more than 0, peaks below will be removed.}

\item{compound_thr}{\code{numeric} specifying intensity threshold (100\% BPI) for peaks in all compound's chemical spectra. If set to more than 0, peaks below will be removed.}
}
\value{
Function returns generated database table and writes it to a csv file in the selected directory (out_dir).
}
\description{
Function generates a chemical reference database table from rda files in the selected directory (rda_dir).
}
\details{
Each rda file must be written for each chemical compond separately and contain a \code{chem.file} list, comprising the following four entries:
\itemize{
\item \code{header} listing experimental details.
\item \code{id} specifying chemical compound identity.
\item \code{reference}
\item \code{analytical} listing all detected peaks, which were built into chemical spectra (i.e. peak groups).
}

The database table contains columns: "peakid", "mz", "rt", "into", "chemid", "dbid", "dbname".
}
\seealso{
\code{\link{annotateDS}}
}
