% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-checkSPIKED.R
\name{checkSPIKED}
\alias{checkSPIKED}
\title{Function to find and plot peaks corresponding to spiked reference compounds}
\usage{
checkSPIKED(fname, pks, raw, spk, out_dir, paramCWT, add_xcmsparams = T,
  rt_err = 3, mz_err = 0.005)
}
\arguments{
\item{fname}{A \code{character} specifying filename of the spiked sample.}

\item{pks}{A \code{DataFrame} class peak table, created for the spiked sample by \emph{pickPEAKS} function.}

\item{raw}{A \code{OnDiskMSnExp} class object for the spiked sample's spectrum, created by the \emph{MSnbase::readMSData} function.}

\item{spk}{A \code{DataFrame} containing information about spiked reference compounds, each row is a unique compound.}

\item{out_dir}{A \code{character} specifying directory where output data will be saved.}

\item{paramCWT}{A \code{CentWaveParam} class object with \emph{centWave} parameters that were used to peak-pick the sample of interest. Object can be created by the \emph{xcms::CentWaveParam} function.}

\item{add_xcmsparams}{A \code{logical} whether \emph{centWave} parameters should be added to the plot.}

\item{rt_err}{A \code{numeric} specifying the window for peak matching in the RT dimension. Default set to 3 (sec).}

\item{mz_err}{A \code{numeric} specifying the window for peak matching in the MZ dimension. Default set to 0.005.}
}
\value{
Function returns a \code{DataFrame} with detected peaks for every spiked compound, also plots spectrum for each peak.
}
\description{
Function to find and plot peaks corresponding to spiked reference compounds
}
