% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanPEAKS}
\alias{cleanPEAKS}
\title{Clean peak table from duplicating peaks}
\usage{
cleanPEAKS(rn, dt_unique, dt)
}
\arguments{
\item{rn}{\code{numeric} specifying a single row number (rowname) of the peak table.}

\item{dt_unique}{\code{data.frame} with unique peaks' \emph{m/z} and \emph{rt} values, with as many rows, as there are unique peaks.}

\item{dt}{\code{data.frame} object containing the peak table from which duplicated peaks should be removed.}
}
\value{
Function returns a single peak entry for the unique \emph{m/z} and \emph{rt} combination.
}
\description{
Return a single entry for every duplicated peak (i.e. peak with the same emph{m/z} and \emph{rt} values).
}
