% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupPEAKS_helpers.R
\name{extractEIC}
\alias{extractEIC}
\title{Obtain extracted ion chromatograms for all picked peaks}
\usage{
extractEIC(raw, pks)
}
\arguments{
\item{raw}{\code{OnDiskMSnExp} class object for the single datafile of interest.}

\item{pks}{\code{DataFrame} object generated with function \code{pickPEAKS}.}
}
\value{
Function returns a \code{list} with an EIC for eack peak in the \code{pks} table.
Individual peaks' EICs can be plotted using base graphics.
}
\description{
Obtain extracted ion chromatograms (EIC) for all picked peaks, provided by \code{pickPEAKS} function.
}
