% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\docType{methods}
\name{filepath,massFlowTemplate-method}
\alias{filepath,massFlowTemplate-method}
\title{Get the the absolute path to the experimental details file, which was used to build the template}
\usage{
\S4method{filepath}{massFlowTemplate}(object)
}
\arguments{
\item{A}{\code{massFlowTemplate} class object}
}
\description{
Get the the absolute path to the experimental details file, which was used to build the template
}
