% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\docType{methods}
\name{fillPEAKS,massFlowTemplate-method}
\alias{fillPEAKS,massFlowTemplate-method}
\alias{fillPEAKS}
\title{Fill peaks}
\usage{
\S4method{fillPEAKS}{massFlowTemplate}(object, fill_value = "into",
  out_dir = NULL, ncores = 2)
}
\arguments{
\item{object}{\code{massFlowTemplate} class object.}

\item{fill_value}{\code{character} specifying which intensity value should be filled and returned, default set to 'into'.}

\item{out_dir}{\code{character} specifying desired directory for output.}

\item{ncores}{\code{numeric} defining number of cores to use for parallelisation. Default set to 2 for parallel implementation.}
}
\value{
Method returns peak table with filled intensities for missed peaks.
}
\description{
Fill peaks
}
\seealso{
\code{\link{loadALIGNED}}, \code{\link{alignPEAKS}}, \code{\link{validPEAKS}}, \code{\link{massFlowTemplate-class}}
}
