% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupCOMPS.R
\name{groupCOMPS}
\alias{groupCOMPS}
\title{Group components across datafiles.}
\usage{
groupCOMPS(files, mz_err = 0.01, rt_err = 0.2, bins = 0.01)
}
\arguments{
\item{bins}{}
}
\description{
Group components across datafiles.
}
\examples{

}
