% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupPEAKS.R
\name{groupPEAKS}
\alias{groupPEAKS}
\title{Get peak-groups representing different chemical spectra in each LC-MS datafile}
\usage{
groupPEAKS(file = NULL, out_dir = NULL, cwt = NULL, ncores = 2,
  thr = 0.95)
}
\arguments{
\item{file}{\code{character} with path to metadata csv file, which must incluce columns 'filename', 'run_order' and 'raw_filepath'.}

\item{out_dir}{\code{character} specifying desired directory for output.}

\item{cwt}{\code{CentWaveParam} class object with parameters for centWave-based peak-picking.}

\item{ncores}{\code{numeric} defining number of cores to use for parallelisation. Default set to 1 for serial implementation.}

\item{thr}{\code{numeric} defining correlation coefficient threshold, above which peak pairs will be considered as correlated. Default set to 0.95.}
}
\value{
For each LC-MS file, function writes a table with picked-peaks and their peak-groups into separate files, in the defined \code{out_dir} directory.
Inputed metadata file is updated to include a column 'proc_filepath', specifying full path to processed csv files peak-group tables.
}
\description{
Function performs peak-picking and peak-grouping for each LC-MS datafile independently.
}
\details{
Function performs peak-picking using the \emph{centWave} algorithm from package \emph{xcms}.
Picked-peaks are then grouped into chemical spectra. For each peak in the sample:
\itemize{
\item Co-eluting peaks are found.
\item EIC correlation between all co-eluting peaks is performed.
\item Network of peaks with high EIC correlation (with coefficients above the selected threshold) is built.
}
Co-eluting peaks within a network of high EIC correlation originate from the same chemical compound and therefore form a chemical spectrum.
}
