% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupPEAKS.R
\name{groupPEAKS_paral}
\alias{groupPEAKS_paral}
\title{Perform peak-picking and grouping into Extracted Chemical Spectra for a single file}
\usage{
groupPEAKS_paral(f, out_dir, cwt, thr, ...)
}
\arguments{
\item{f}{\code{character} specifying absolute path to a single mzML/CDF file.}

\item{out_dir}{\code{character} specifying absolute path to directory for output.}

\item{cwt}{\code{CentWaveParam} class object with parameters for peak-picking. Object can be created by the \emph{xcms::CentWaveParam} function.}

\item{thr}{\code{numeric} defining Pearson correlation coefficient threshold, above which peak pairs will be considered as correlated.}
}
\value{

}
\description{
Function enabes parallel or serial implementation of peak-picking and grouping.
This function is called from inside \code{\link{groupPEAKS}}, for each sample in the experiment separately.
}
