% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupPEAKSspec.R
\name{groupPEAKSspec}
\alias{groupPEAKSspec}
\title{Group peaks of a single LC-MS spectrum using EIC correlation}
\usage{
groupPEAKSspec(pks, eic, out_dir, fname, thr = thr, return = FALSE)
}
\arguments{
\item{pks}{\code{DataFrame} containing the peak table, created by \emph{pickPEAKS} function.}

\item{eic}{\code{list} containing extracted ion chromatograms for each peak in the \code{pks} table.}

\item{out_dir}{\code{character} specifying directory where output data will be saved.}

\item{fname}{\code{character} specifying LC-MS filename.}

\item{thr}{\code{numeric} defining Pearson correlation coefficient threshold, above which peak pairs will be considered as correlated.}

\item{return}{\code{logical} whether to return the obtained peak-group table. Set to TRUE if working with function interactivily.}
}
\value{
Function returns an updated peak table.
Column \code{"peakgr"} contains peak-group ID for each peak.
Peaks not correlated to any of its co-eluting peaks are removed from the table.
Function also writes updated peak table in the specified directory.
}
\description{
Function builds peak-groups of co-eluting peaks that are correlated to each other in a single LC-MS spectrum.
}
