% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{loadALIGNED}
\alias{loadALIGNED}
\title{Build sample alignment and annotation template using already aligned samples}
\usage{
loadALIGNED(file, db = NULL, mz_err = 0.01, rt_err = 2,
  bins = 0.01)
}
\arguments{
\item{file}{A \code{character} with path to the csv file, specifying samples filenames and their acquisition order.}

\item{db}{\code{NULL}, or a \code{massFlowDB} class object, built with \code{massFlowDB()} function.}

\item{mz_err}{A \code{numeric} specifying the window for peak matching in the MZ dimension. Default set to 0.01.}

\item{rt_err}{A \code{numeric} specifying the window for peak matching in the RT dimension. Default set to 2 (sec).}

\item{bins}{A \code{numeric} defying step size used in component's spectra binning and vector generation. Step size represents MZ dimension (default set to 0.01).}
}
\value{
A \code{massFlowTemplate} class object.
}
\description{
Function handles the construction of \code{massFlowTemplate} class object from already aligned samples.
Function enables user to continue an interrupted peak alignment process, facilitated by \code{alignPEAKS} function.
}
\details{
Arguments are identical to the ones used by \code{\link{buildTMP}} constructor function.
}
\seealso{
\code{\link{massFlowTemplate}} class.
}
