% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods-massFlowDB.R
\docType{class}
\name{massFlowDB-class}
\alias{massFlowDB-class}
\alias{massFlowDB}
\alias{show,massFlowDB-method}
\alias{show}
\alias{filepath,massFlowDB-method}
\alias{filepath}
\alias{chemicals,massFlowDB-method}
\alias{chemicals}
\title{Chemical reference database template}
\usage{
\S4method{show}{massFlowDB}(object)

\S4method{filepath}{massFlowDB}(object)

\S4method{chemicals}{massFlowDB}(object)
}
\description{
\code{massFlowDB} class encapsulates information provided in the chemical reference database.
}
\details{
Chemical reference database file lists peaks \code{centWave}-detected in the reference LC-MS datafiles.#' 
The database file is a single csv file, which must contain columns: "peakid", "mz", "rt", "into", "peakgr", "chemid", "dbid", "dbname".

\code{massFlowDB} object will be used to initialise \code{\link{massFlowTemplate-class}} object.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{\code{character} specifying path to database csv file.}

\item{\code{db}}{\code{data.frame} containing database template.}
}}

