% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{massFlowDB-class}
\alias{massFlowDB-class}
\title{Chemical reference database template.}
\description{
\code{massFlowDB} class encapsulates information provided in the chemical reference database.
Database contains all of the detected peaks for every chemical in the database. Peaks must be grouped into spectra using \code{pickPEAKS()} function prior to building of the \code{massFlowDB} class object.
The final database file is a single csv file, which must contain columns: "peakid", "mz", "rt", "into", "peakgr", "chemid", "dbid", "dbname".
Database will be used to initialise \code{massFlowTemplate} class object.
Database "peakid" and "peakgr" values will retain unmodified in the template, while every sample in the study will be aligned with them.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{A \code{character} specifying path to database csv file.}

\item{\code{db}}{A \code{data.frame} containing database template. Template must contain}
}}

