% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{massFlowTemplate-class}
\alias{massFlowTemplate-class}
\title{Sample alignment and annotation template.}
\description{
Sample alignment and annotation template.
}
\details{
\code{massFlowTemplate} object stores the sample alignment and annotation template.
Template is initiated using the first datafile in the study and (if provided) the chemical reference database (\code{massFlowDB} object).
To initiate the template, constructor function \code{\link{massFlowTemplate}} must be called, providing a "filepath" to the csv file with the details on the study samples. This csv file must contain columns "filepaths" and "run_order".
Each study sample must be processed with \code{\link{groupPEAKS}} function first in order to obtain spectral peak groups.
With every round of sample alignment & annotation, template is updated and an intermediate sample alignment output is written as a csv file.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{A \code{character} specifying the absolute path to a \code{csv} file with study sample names and their acquisition (run) order.}

\item{\code{db_filepath}}{A \code{character} specifying the absolute path to the database \code{csv} file (if was used in the template built).}

\item{\code{samples}}{A \code{data.frame} storing study sample names and their acquisition order.}

\item{\code{tmp}}{A \code{data.frame} storing sample alignment and annotation template.}

\item{\code{data}}{A \code{list} containg annotated and aligned peak tables for the already processed samples.}
}}

