% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/methods-massFlowTemplate.R
\docType{class}
\name{massFlowTemplate-class}
\alias{massFlowTemplate-class}
\alias{massFlowTemplate}
\alias{show,massFlowTemplate-method}
\alias{show}
\alias{filepath,massFlowTemplate-method}
\alias{filepath}
\title{massFlowTemplate, a class for sample alignment and annotation}
\usage{
\S4method{show}{massFlowTemplate}(object)

\S4method{filepath}{massFlowTemplate}(object)
}
\description{
This class enables the alignment and annotation of LC-MS samples.
It takes files containing groups of structurally related peaks, generated by \code{\link{groupPEAKS}} function, and aligns them in their original sample acquisition order.
It validates peak groups obtained during alignment by correlating peak intensities in all samples.
Finally, it fills in missing peak intensity values using raw data.
}
\details{
\code{massFlowTemplate} object stores sample alignment and annotation data.

\itemize{
\item Slot @tmp is the template, listing information for every detected peak, such as the \emph{m/z}, \emph{rt}, \emph{into} values, the assigned peak-group, as well as chemical identifications (if any).
\item Template is first built using the first datafile in the study, and, if available, the chemical reference database (\code{\link{massFlowDB}} object).
To initiate template built, constructor function \code{\link{buildTMP}} must be called, providing a "filepath" to the csv file with the details on the study samples.
\item Template is updated with every round of sample alignment.
Alignment results are stored in slot @data for each sample separately.
}
}
\section{Slots}{

\describe{
\item{\code{filepath}}{\code{character} specifying the absolute path to a \code{csv} file with study sample names and their acquisition (run) order.}

\item{\code{db_filepath}}{\code{character} specifying the absolute path to the database \code{csv} file (if was used in the template built).}

\item{\code{samples}}{\code{data.frame} storing study sample names and their acquisition order.}

\item{\code{tmp}}{\code{data.frame} storing sample alignment and annotation template.}

\item{\code{data}}{\code{list} containg annotated and aligned peak tables for the already processed samples.}

\item{\code{params}}{\code{list} containg alignment and annotation parameters.}
}}

\seealso{
\code{\link{buildTMP}}, \code{\link{alignPEAKS}}
}
