% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{massFlowTemplate-class}
\alias{massFlowTemplate-class}
\title{Class massFlowTemplate, a class for sample alignment and annotation}
\description{
This class enables the alignment and annotation of LC-MS samples.
It takes files containing groups of structurally related peaks, generated by \code{groupPEAKS} function, and aligns them in their original sample acquisition order.
It validates peak groups obtained during alignment by correlating peak intensities in all samples.
Finally, it fills in missing peak intensity values using raw data.
}
\details{
\code{massFlowTemplate} object stores sample alignment and annotation data.

Slot \code{tmp} is the template, listing information for every detected peak, such as the \emph{mz}, \emph{rt}, \emph{into} values, the assigned peak-group, as well as chemical identifications (if any).
Template is initiated using the first datafile in the study and the chemical reference database (\code{massFlowDB} object), if available.
To initiate template built, function \code{buildTMP} must be called, providing a "filepath" to the csv file with the details on the study samples.

Template is updated with every round of sample alignment. Alignment results are stored in slot \code{data} for each sample separately.
}
\section{Slots}{

\describe{
\item{\code{filepath}}{A \code{character} specifying the absolute path to a \code{csv} file with study sample names and their acquisition (run) order.}

\item{\code{db_filepath}}{A \code{character} specifying the absolute path to the database \code{csv} file (if was used in the template built).}

\item{\code{samples}}{A \code{data.frame} storing study sample names and their acquisition order.}

\item{\code{tmp}}{A \code{data.frame} storing sample alignment and annotation template.}

\item{\code{data}}{A \code{list} containg annotated and aligned peak tables for the already processed samples.}

\item{\code{params}}{A \code{list} containg alignment and annotation parameters.}
}}

