% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-groupPEAKS_helpers.R
\name{pickPEAKS}
\alias{pickPEAKS}
\title{Peak detection using the centWave method}
\usage{
pickPEAKS(raw, cwt, fname)
}
\arguments{
\item{raw}{\code{OnDiskMSnExp} class object for the single LC-MS spectrum of interest. Such object can be created by the \emph{MSnbase::readMSData} function.}

\item{cwt}{\code{CentWaveParam} class object with parameters for peak-picking. Object can be created by the \emph{xcms::CentWaveParam} function.}

\item{fname}{\code{character} specifying datafile name.}
}
\value{
Function returns a \code{DataFrame} object with picked-peaks and centWave details.
}
\description{
High resolution, centroid LC/MS datafiles are peak-picked using centWave algorithm, implemented by \emph{xcms} package [Tautenhahn 2008].
\code{pickPEAKS} is applied to a single datafile.
}
