% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-alignPEAKS_helpers.R
\name{scaleSPEC}
\alias{scaleSPEC}
\title{Scale a vector representing peaks' intensity values}
\usage{
scaleSPEC(spec, m = 0.6, n = 3)
}
\arguments{
\item{spec}{\code{data.frame} with columns 'into' and 'mz'.}

\item{m}{\code{numeric}, set to 0.6 by default.}

\item{n}{\code{numeric}, set to 3 by default.}
}
\value{
Function returns a \code{numeric} vector with scaled intensity values.
}
\description{
Function scales a vector representing peaks' intensity values before dot-product estimation of two vectors.
Currently, scaling according to method in Stein & Scott, 1994 is supported.
}
\seealso{
\code{\link{getCOSmat}}
}
