% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-massFlowTemplate.R
\docType{methods}
\name{validPEAKS,massFlowTemplate-method}
\alias{validPEAKS,massFlowTemplate-method}
\alias{validPEAKS}
\title{Validate aligned peaks and their corresponding peak-groups}
\usage{
\S4method{validPEAKS}{massFlowTemplate}(object, ncores = 1, out_dir)
}
\arguments{
\item{object}{\code{massFlowTemplate} class object.}

\item{ncores}{\code{numeric} defining number of cores to use for parallelisation. Default set to 1 for serial implementation.}

\item{out_dir}{\code{character} specifying desired directory for output.}
}
\value{
Method returns validated peak-groups.
}
\description{
Validate aligned peaks and their corresponding peak-groups
}
